+++ packages/snap-networks-utils/src/dedupe/index.ts
@@ -1,0 +1,1 @@
+export { InFlightCoalescer } from './InFlightCoalescer';